working on memory management code

+++ core/vm/arch/win32/memory.h
@@ -52,7 +52,7 @@
 #define CACHE_SIZE -4
 
 struct StackOperMemory {
-  long* op_stack;
+  size_t* op_stack;
   long* stack_pos;
 };
 
@@ -71,8 +71,8 @@ struct CollectionInfo {
 };
 
 struct ClassMethodId {
-  long* self;
-  long* mem;
+  size_t* self;
+  size_t* mem;
   long cls_id;
   long mthd_id;
 };
@@ -82,10 +82,10 @@ class MemoryManager {
   static StackProgram* prgm;
   
   // reference containers
-  static unordered_map<long*, ClassMethodId*> jit_roots;
+  static unordered_map<size_t*, ClassMethodId*> jit_roots;
   static unordered_set<StackFrame**> pda_frames;
   static unordered_map<StackFrameMonitor*, StackFrameMonitor*> pda_monitors; // deleted elsewhere
-  static vector<long*> allocated_memory;
+  static vector<size_t*> allocated_memory;
 
   // memory caches
   static stack<char*> cache_pool_16;
@@ -108,10 +108,10 @@ class MemoryManager {
   static long collected_count;
 
   // if return true, trace memory otherwise do not
-  static inline bool MarkMemory(long* mem);
-  static inline bool MarkValidMemory(long* mem);
+  static inline bool MarkMemory(size_t* mem);
+  static inline bool MarkValidMemory(size_t* mem);
 
-  static inline StackClass* GetClassMapping(long* mem) {
+  static inline StackClass* GetClassMapping(size_t* mem) {
 #ifndef _GC_SERIAL
       EnterCriticalSection(&allocated_cs);
 #endif
@@ -135,7 +135,7 @@ public:
 
   // free static resources
   static void Clear() {
-    unordered_map<long*, ClassMethodId*>::iterator id_iter;
+    unordered_map<size_t*, ClassMethodId*>::iterator id_iter;
     for(id_iter = jit_roots.begin(); id_iter != jit_roots.end(); ++id_iter) {
       ClassMethodId* tmp = id_iter->second;
       // delete
@@ -144,7 +144,7 @@ public:
     }
 
     while(!allocated_memory.empty()) {
-      long* temp = allocated_memory.front();
+      size_t* temp = allocated_memory.front();
       allocated_memory.erase(allocated_memory.begin());      
       temp -= EXTRA_BUF_SIZE;
       free(temp);
@@ -205,8 +205,8 @@ public:
   static uintptr_t WINAPI CheckPdaRoots(LPVOID arg);
   
   // add and remove jit roots
-  static void AddJitMethodRoot(long cls_id, long mthd_id, long* self, long* mem, long offset);
-  static void RemoveJitMethodRoot(long* mem);
+  static void AddJitMethodRoot(long cls_id, long mthd_id, size_t* self, size_t* mem, long offset);
+  static void RemoveJitMethodRoot(size_t* mem);
 
   // add and remove pda roots
   static void AddPdaMethodRoot(StackFrame** frame);
@@ -215,13 +215,13 @@ public:
   static void RemovePdaMethodRoot(StackFrameMonitor* monitor);
   
   // sweeps memory
-  static void CheckMemory(long* mem, StackDclr** dclrs, const long dcls_size, const long depth);
+  static void CheckMemory(size_t* mem, StackDclr** dclrs, const long dcls_size, const long depth);
 
   // sweeps an object and it's references
-  static void CheckObject(long* mem, bool is_obj, const long depth);
+  static void CheckObject(size_t* mem, bool is_obj, const long depth);
 
   // memory allocation
-  static long* AllocateObject(const wchar_t* obj_name, size_t* op_stack, long stack_pos, bool collect = true) {
+  static size_t* AllocateObject(const wchar_t* obj_name, size_t* op_stack, long stack_pos, bool collect = true) {
     StackClass* cls = prgm->GetClass(obj_name);
     if(cls) {
       return AllocateObject(cls->GetId(), op_stack, stack_pos, collect);
@@ -233,18 +233,18 @@ public:
   //
   // object and array allocation
   //
-  static long* AllocateObject(const long obj_id, size_t* op_stack, long stack_pos, bool collect = true);
-  static long* AllocateArray(const long size, const MemoryType type, size_t* op_stack, long stack_pos, bool collect = true);
+  static size_t* AllocateObject(const long obj_id, size_t* op_stack, long stack_pos, bool collect = true);
+  static size_t* AllocateArray(const long size, const MemoryType type, size_t* op_stack, long stack_pos, bool collect = true);
 
   //
   // object verification
   //
-  static long* ValidObjectCast(long* mem, long to_id, int* cls_hierarchy, int** cls_interfaces);
+  static size_t* ValidObjectCast(size_t* mem, long to_id, int* cls_hierarchy, int** cls_interfaces);
   
   //
   // returns the class reference for an object instance
   //
-  static inline StackClass* GetClass(long* mem) {
+  static inline StackClass* GetClass(size_t* mem) {
     if(mem && mem[TYPE] == MemoryType::NIL_TYPE) {
       return (StackClass*)mem[SIZE_OR_CLS];
     }
@@ -255,7 +255,7 @@ public:
   //
   // returns a unique object id for an instance
   //
-  static inline long GetObjectID(long* mem) {
+  static inline long GetObjectID(size_t* mem) {
     StackClass* klass = GetClass(mem);
     if(klass) {
       return klass->GetId();
